requirements.txt
# gcp SDKs
google-cloud-billing>=1.12.0
google-cloud-asset>=3.19.0
google-cloud-compute>=1.15.0

# authentication
google-auth>=2.23.0

# configuration management
PyYAML>=6.0

# testing
pytest>=7.4.0
pytest-cov>=4.1.0

# code quality
black>=23.0.0
flake8>=6.1.0
